+++ is_prime_prog/is_prime.hpp
@@ -1,4 +1,4 @@
 #include <iostream>
 #include <cmath>
 
-std::string is_prime(int val);
+std::string is_prime(long long val);
